Added styled homepage and added lightbox effect to create a pet dialogue form
