+++ src/app/mypets/page.tsx
@@ -1,6 +1,9 @@
 "use client";
 import { useState, Dispatch, SetStateAction } from "react";
 import { PetInputForm } from "@/components/PetInputForm";
+import { TopBar } from "@/components/TopBar";
+import { PetInformation } from "@/components/PetInformation";
+import { AddVaxInfoForm } from "@/components/AddVaxInfoForm";
 
 //Dummy data to build out the UI without fetching
 const pets = [
@@ -48,71 +51,9 @@ const pets = [
   },
 ];
 
-export default function MyPets() {
-  const [openForm, setOpenForm] = useState<boolean>(false);
-  const [petDisplayIndex, setPetDisplayIndex] = useState<number>(-1);
-
-  const [displayvaxForms, setDisplayVaxForms] = useState<boolean>(false);
-
-  if (petDisplayIndex == -1) {
-    return (
-      <div className="flex flex-col place-content-center min-w-full max-w-screen-2xl">
-        <div className="flex flex-row justify-between relative">
-          <h1 className="text-7xl">My Pets</h1>
-          <div className="absolute bottom-1 right-0">
-            {pets.length ? (
-              <AddPetsButton onClickHandler={(e) => setOpenForm(!openForm)} />
-            ) : null}
-          </div>
-        </div>
-        <dialog className="border-2 z-10" open={openForm}>
-          <div className="relative">
-            <button
-              onClick={() => setOpenForm(false)}
-              className="absolute top-1 right-2"
-            >
-              x
-            </button>
-            <PetInputForm />
-          </div>
-        </dialog>
-        {displayvaxForms && (
-          <dialog className="border-2 z-30" open={displayvaxForms}>
-            <AddVaxInfoForm setDisplayVaxForms={setDisplayVaxForms} />
-          </dialog>
-        )}
-        <div className="">
-          <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4 mt-2">
-            {pets.length ? (
-              pets.map((pet, index) => (
-                <PetDisplayTile
-                  key={pet.name}
-                  {...pet}
-                  index={index}
-                  setPetDisplayIndex={setPetDisplayIndex}
-                  setDisplayVaxForms={setDisplayVaxForms}
-                />
-              ))
-            ) : (
-              <div>
-                You have no pets currently. Add one here.
-                <AddPetsButton onClickHandler={(e) => setOpenForm(!openForm)} />
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-    );
-  } else {
-    return (
-      <PetInformation
-        {...pets[petDisplayIndex]}
-        setPetDisplayIndex={setPetDisplayIndex}
-      />
-    );
-  }
-}
-
+type AddPetsButtonProps = {
+  onClickHandler: React.MouseEventHandler<HTMLButtonElement>;
+};
 type PetDisplayTileProps = {
   name: string;
   photo: string;
@@ -127,6 +68,87 @@ type PetDisplayTileProps = {
   setDisplayVaxForms: Dispatch<SetStateAction<boolean>>;
   setPetDisplayIndex: Dispatch<SetStateAction<number>>;
 };
+
+export default function MyPets() {
+  const [openForm, setOpenForm] = useState<boolean>(false);
+  const [petDisplayIndex, setPetDisplayIndex] = useState<number>(-1);
+
+  const [displayvaxForms, setDisplayVaxForms] = useState<boolean>(false);
+
+  if (petDisplayIndex == -1) {
+    return (
+      <div className="flex flex-col flex-grow place-content-center min-w-full max-w-2xl mx-auto">
+        <TopBar />
+        <div className="flex flex-col flex-grow justify-center max-w-screen-xl mx-auto">
+          <div className="flex flex-row justify-between relative">
+            <h1 className="text-7xl">My Pets</h1>
+            <div className="absolute bottom-1 right-0">
+              {pets.length ? (
+                <AddPetsButton onClickHandler={(e) => setOpenForm(!openForm)} />
+              ) : null}
+            </div>
+          </div>
+          <div
+            className={
+              openForm
+                ? "fixed inset-0 bg-black bg-opacity-40 flex justify-center items-center z-50"
+                : "hidden"
+            }
+          >
+            <dialog
+              className="z-10 drop-shadow-2xl rounded-lg dialog-animation w-full md:w-auto"
+              open={openForm}
+            >
+              <div className="relative">
+                <button
+                  onClick={() => setOpenForm(false)}
+                  className="absolute top-1 right-2"
+                >
+                  x
+                </button>
+                <PetInputForm />
+              </div>
+            </dialog>
+          </div>
+          {displayvaxForms && (
+            <dialog className="border-2 z-30" open={displayvaxForms}>
+              <AddVaxInfoForm setDisplayVaxForms={setDisplayVaxForms} />
+            </dialog>
+          )}
+          <div className="">
+            <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4 mt-2">
+              {pets.length ? (
+                pets.map((pet, index) => (
+                  <PetDisplayTile
+                    key={pet.name}
+                    {...pet}
+                    index={index}
+                    setPetDisplayIndex={setPetDisplayIndex}
+                    setDisplayVaxForms={setDisplayVaxForms}
+                  />
+                ))
+              ) : (
+                <div>
+                  You have no pets currently. Add one here.
+                  <AddPetsButton
+                    onClickHandler={(e) => setOpenForm(!openForm)}
+                  />
+                </div>
+              )}
+            </div>
+          </div>
+        </div>
+      </div>
+    );
+  } else {
+    return (
+      <PetInformation
+        {...pets[petDisplayIndex]}
+        setPetDisplayIndex={setPetDisplayIndex}
+      />
+    );
+  }
+}
 
 function PetDisplayTile({
   name,
@@ -179,10 +201,6 @@ function PetDisplayTile({
   );
 }
 
-type AddPetsButtonProps = {
-  onClickHandler: React.MouseEventHandler<HTMLButtonElement>;
-};
-
 function AddPetsButton({ onClickHandler }: AddPetsButtonProps) {
   return (
     <button
@@ -192,137 +210,4 @@ function AddPetsButton({ onClickHandler }: AddPetsButtonProps) {
       Add Pet
     </button>
   );
-}
-
-type PetInformationProps = {
-  name: string;
-  photo: string;
-  description: string;
-  species: string;
-  breed: string;
-  markings: string[];
-  gender: string;
-  spayed: boolean;
-  microchip: string;
-  setPetDisplayIndex: Dispatch<SetStateAction<number>>;
-};
-
-function PetInformation({
-  name,
-  photo,
-  description,
-  species,
-  breed,
-  markings,
-  gender,
-  spayed,
-  microchip,
-  setPetDisplayIndex,
-}: PetInformationProps) {
-  return (
-    <div className="border-2 flex flex-col items-center p-4 rounded-br-lg rounded-tl-lg shadow-md gap-4 relative">
-      <button
-        onClick={() => setPetDisplayIndex(-1)}
-        className="absolute top-0 right-2"
-      >
-        x
-      </button>
-      <div className="">
-        <img
-          className="object-cover rounded-full w-32 h-32 object-cover rounded-full"
-          src={photo}
-        />
-      </div>
-      <h2 className="text-5xl">{name}</h2>
-      {microchip && (
-        <p className="text-center leading-8">
-          Chip Number <br />
-          <span className="border-2 py-1 px-3 rounded-full shadow-md border-indigo-500">
-            {microchip}
-          </span>
-        </p>
-      )}
-      <p className="font-light text-center h-12">{description.slice(0, 100)}</p>
-      <div className="flex flex-row gap-2">
-        <p>Markings: </p>
-        <div className="flex flex-wrap gap-2">
-          {markings.map((marking) => (
-            <p
-              className="border-2 py-1 px-2 rounded-full shadow-md border-indigo-500 text-xs"
-              key={marking}
-            >
-              {marking}
-            </p>
-          ))}
-        </div>
-      </div>
-      <div className="grid grid-cols-2 lg:grid-cols-4 w-7/8 text-xs pb-10">
-        <p>Species: </p> <p>{species}</p>
-        <p>Breed: </p> <p>{breed}</p>
-        <p>Gender: </p> <p>{gender}</p>
-        <p>Spayed: </p> <p>{spayed}</p>
-      </div>
-    </div>
-  );
-}
-
-type AddVaxInfoFormProps = {
-  setDisplayVaxForms: Dispatch<SetStateAction<boolean>>;
-};
-
-function AddVaxInfoForm({ setDisplayVaxForms }: AddVaxInfoFormProps) {
-  const [petInfo, setPetInfo] = useState([
-    <VaxForm key="VaxForm" />,
-    <TestForm key="TestForm" />,
-    <CertsForm key="CertsForm" />,
-  ]);
-
-  const [currentFormIndex, setCurrentFormIndex] = useState<number>(0);
-
-  const [vaxFormDetails, setVaxFormDetails] = useState();
-  const [testFormDetails, setTestFormDetails] = useState();
-  const [certFormDetails, setCertFormDetails] = useState();
-
-  return (
-    <div className="relative p-4">
-      <button
-        className="absolute top-0 right-1"
-        onClick={() => setDisplayVaxForms(false)}
-      >
-        x
-      </button>
-      <div className="flex flex-row gap-4">
-        <button onClick={() => setCurrentFormIndex(0)}>
-          Submit Vaccination
-        </button>
-        <button onClick={() => setCurrentFormIndex(1)}>Submit Test</button>
-        <button onClick={() => setCurrentFormIndex(2)}>
-          Submit Certification
-        </button>
-      </div>
-      <form>{petInfo[currentFormIndex]}</form>
-    </div>
-  );
-}
-
-type VaxFormProps = {};
-
-function VaxForm({}: VaxFormProps) {
-  return (
-    <>
-      <input type="date" />
-    </>
-  );
-}
-
-type TestFormProps = {};
-
-function TestForm({}: TestFormProps) {
-  return <p>Test Form</p>;
-}
-
-type CertFormProps = {};
-
-function CertsForm({}: CertFormProps) {
-  return <p>Certs Form</p>;
 }
